[ADD] split component for add node
