+++ client/src/app/pages/pipeline/components/pipeline-node/components/add-node/add-node.component.html
@@ -1,10 +1,9 @@
-<app-button (click)="modal.openModal()">add process</app-button>
 <app-modal #ProcessAdd title="add processor">
     <div class="list__wrapper">
         <button
             class="list__button"
             *ngFor="let process of processes | keyvalue"
-            (click)="addProcess(process.key)"
+            (click)="addNodeHandle(nodeData.position, process.value)"
             [title]="process.value.title"
         >
             <app-svg [iconType]="process.value.icon"></app-svg>
