+++ client/src/app/pages/pipeline/components/pipeline-node/components/add-node/add-node.component.scss
