+++ client/src/app/pages/pipeline/pipeline.module.ts
@@ -5,7 +5,7 @@ import {SvgModule} from '../../components/svg/svg.module';
 import {RunButtonComponent} from './components/run-button/run-button.component';
 import {LoadingButtonModule} from '../../components/loading-button/loading-button.module';
 
-import {ProcessListComponent} from './components/process-list/process-list.component';
+import {addNodeComponent} from './components/pipeline-node/components/add-node/add-node.component';
 
 import {ButtonModule} from 'src/app/components/button/button.module';
 import {ModalModule} from 'src/app/components/modal/modal.module';
@@ -16,16 +16,18 @@ import {SettingItemModalComponent} from './components/setting-item-modal/setting
 import {BoardComponent} from './components/board/board.component';
 import {AngularDraggableModule} from 'angular2-draggable';
 import {NgxDraggabillyModule} from 'ngx-draggabilly';
+import { RemoveNodeComponent } from './components/pipeline-node/components/remove-node/remove-node.component';
 
 @NgModule({
     declarations: [
         PipelineComponent,
         RunButtonComponent,
-        ProcessListComponent,
+        addNodeComponent,
         PipelineNodeComponent,
         PipelineBoardComponent,
         SettingItemModalComponent,
         BoardComponent,
+        RemoveNodeComponent,
     ],
 
     imports: [
